backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
fdb==2.0.3
pydantic==2.11.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
